move tests folder up a level
