+++ AlexaSkillFiles-SkiResort/tests/intent-sample-requests/forecast-week.intent.js
