+++ AlexaSkillFiles-SkiResort/tests/intent-sample-requests/forecast-weekday.intent.js
@@ -1,27 +1,36 @@
-const { AlexaAppId } = require('../../secrets/credentials');
+const { AlexaAppId } = require('../../lambda/secrets/credentials');
 
 module.exports = {
 	"version": "1.0",
 	"session": {
-    "new": true,
-    "sessionId": "sessionId123",
-    "application": {
-      "applicationId": AlexaAppId
-    },
-    "attributes": {},
-    "user": {
-      "userId": "userId123"
-    }
-  },
+		"new": true,
+		"sessionId": "sessionId123",
+		"application": {
+			"applicationId": AlexaAppId
+		},
+		"user": {
+			"userId": "userId123"
+		}
+	},
 	"request": {
 		"type": "IntentRequest",
 		"requestId": "requestId123",
-    "locale": "en-US",
-		"timestamp": "2021-09-11T20:27:34Z",
+		"locale": "en-US",
+		"timestamp": "2021-04-08T02:57:13Z",
 		"intent": {
-			"name": "snowReportOvernight",
+			"name": "forecastWeekDay",
 			"confirmationStatus": "NONE",
 			"slots": {
+				"Day": {
+					"name": "Day",
+					"value": "friday",
+					"confirmationStatus": "NONE",
+					"source": "USER",
+					"slotValue": {
+						"type": "Simple",
+						"value": "friday"
+					}
+				},
 				"Resort": {
 					"name": "Resort",
 					"value": "Stevens pass",
@@ -29,7 +38,7 @@ module.exports = {
 						"resolutionsPerAuthority": [
 							{
 								"authority": "authorityID.LIST_OF_RESORTS",
-                "status": {
+								"status": {
 									"code": "ER_SUCCESS_MATCH"
 								},
 								"values": [
@@ -52,7 +61,7 @@ module.exports = {
 							"resolutionsPerAuthority": [
 								{
 									"authority": "authorityID.LIST_OF_RESORTS",
-                "status": {
+									"status": {
 										"code": "ER_SUCCESS_MATCH"
 									},
 									"values": [
