+++ AlexaSkillFiles-SkiResort/tests/intent-sample-requests/forecastToday/forecast-today.intent.js
