+++ AlexaSkillFiles-SkiResort/tests/intent-sample-requests/new-session/session-start.intent.js
