+++ AlexaSkillFiles-SkiResort/tests/intent-sample-requests/snow-report-season-total.intent.js
@@ -1,36 +1,27 @@
-const { AlexaAppId } = require('../../secrets/credentials');
+const { AlexaAppId } = require('../../lambda/secrets/credentials');
 
 module.exports = {
 	"version": "1.0",
 	"session": {
 		"new": true,
 		"sessionId": "sessionId123",
-		"application": {
-			"applicationId": AlexaAppId
-		},
-		"user": {
-			"userId": "userId123"
-		}
+    "application": {
+      "applicationId": AlexaAppId
+    },
+    "attributes": {},
+    "user": {
+      "userId": "userId123"
+    }
 	},
 	"request": {
 		"type": "IntentRequest",
 		"requestId": "requestId123",
-		"locale": "en-US",
-		"timestamp": "2021-05-12T04:29:24Z",
+    "locale": "en-US",
+		"timestamp": "2021-09-11T20:25:11Z",
 		"intent": {
-			"name": "temperatureWeekDay",
+			"name": "snowReportSeasonTotal",
 			"confirmationStatus": "NONE",
 			"slots": {
-				"Day": {
-					"name": "Day",
-					"value": "friday",
-					"confirmationStatus": "NONE",
-					"source": "USER",
-					"slotValue": {
-						"type": "Simple",
-						"value": "friday"
-					}
-				},
 				"Resort": {
 					"name": "Resort",
 					"value": "Stevens pass",
@@ -38,7 +29,7 @@ module.exports = {
 						"resolutionsPerAuthority": [
 							{
 								"authority": "authorityID.LIST_OF_RESORTS",
-								"status": {
+                "status": {
 									"code": "ER_SUCCESS_MATCH"
 								},
 								"values": [
@@ -61,7 +52,7 @@ module.exports = {
 							"resolutionsPerAuthority": [
 								{
 									"authority": "authorityID.LIST_OF_RESORTS",
-									"status": {
+                "status": {
 										"code": "ER_SUCCESS_MATCH"
 									},
 									"values": [
