+++ AlexaSkillFiles-SkiResort/tests/intent-sample-requests/temperature-tonight.intent.js
@@ -1,25 +1,24 @@
-const { AlexaAppId } = require('../../secrets/credentials');
+const { AlexaAppId } = require('../../lambda/secrets/credentials');
 
 module.exports = {
 	"version": "1.0",
 	"session": {
 		"new": true,
 		"sessionId": "sessionId123",
-    "application": {
-      "applicationId": AlexaAppId
-    },
-    "attributes": {},
-    "user": {
-      "userId": "userId123"
-    }
+		"application": {
+			"applicationId": AlexaAppId
+		},
+		"user": {
+			"userId": "userId123"
+		}
 	},
 	"request": {
 		"type": "IntentRequest",
 		"requestId": "requestId123",
-    "locale": "en-US",
-		"timestamp": "2021-09-11T20:25:11Z",
+		"locale": "en-US",
+		"timestamp": "2021-05-04T04:49:23Z",
 		"intent": {
-			"name": "snowReportSeasonTotal",
+			"name": "temperatureTonight",
 			"confirmationStatus": "NONE",
 			"slots": {
 				"Resort": {
@@ -29,7 +28,7 @@ module.exports = {
 						"resolutionsPerAuthority": [
 							{
 								"authority": "authorityID.LIST_OF_RESORTS",
-                "status": {
+								"status": {
 									"code": "ER_SUCCESS_MATCH"
 								},
 								"values": [
@@ -52,7 +51,7 @@ module.exports = {
 							"resolutionsPerAuthority": [
 								{
 									"authority": "authorityID.LIST_OF_RESORTS",
-                "status": {
+									"status": {
 										"code": "ER_SUCCESS_MATCH"
 									},
 									"values": [
@@ -71,4 +70,4 @@ module.exports = {
 			}
 		}
 	}
-}
+};
